Refactor: 부서 관련 DTO 패키지 분리 및 CursorPageResponseDto 타입 수정 [#45]
