+++ src/main/java/com/hrbank3/hrbank3/dto/CursorPageResponseDto.java
@@ -1,10 +1,10 @@
-package com.hrbank.hrbank.dto;
+package com.hrbank3.hrbank3.dto;
 
 import java.util.List;
 
 public record CursorPageResponseDto<T>(
         List<T> content,      // 실제 데이터 목록
-        Long nextCursor,      // 다음 페이지 시작 기준 ID
+        String nextCursor,      // 다음 페이지 시작 기준 ID
         Long nextIdAfter,     // 다음 페이지 시작 ID
         int size,             // 현재 페이지 데이터 개수
         long totalElements,   // 전체 데이터 개수
