+++ src/main/java/com/hrbank3/hrbank3/repository/DepartmentRepositoryCustom.java
@@ -1,7 +1,7 @@
-package com.hrbank.hrbank.repository;
+package com.hrbank3.hrbank3.repository;
 
-import com.hrbank.hrbank.dto.CursorPageResponseDto;
-import com.hrbank.hrbank.dto.DepartmentDto;
+import com.hrbank3.hrbank3.dto.CursorPageResponseDto;
+import com.hrbank3.hrbank3.dto.department.DepartmentDto;
 
 public interface DepartmentRepositoryCustom {
