+++ src/test/java/com/hrbank/hrbank/HrbankApplicationTests.java
@@ -1,4 +1,4 @@
-package com.hrbank.hrbank;
+package com.hrbank3.hrbank3;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
